todos2/routers/admin.py
from fastapi import APIRouter, Depends, HTTPException, status, Path
from typing import Annotated
from sqlalchemy.orm import Session
from ..models import Todos
from ..database4 import sessionlocal
from .auth import get_current_user

router = APIRouter(
    prefix='/admin',
    tags=['/admin']
)


def get_db():
    db = sessionlocal()
    try:
        yield db
    finally:
        db.close()


db_dependency = Annotated[Session, Depends(get_db)]
user_dependency = Annotated[dict, Depends(get_current_user)]


@router.get("/", status_code=status.HTTP_200_OK)
async def read_all(user: user_dependency, db: db_dependency):
    if user is None or user.get('user_role') != 'admin':
        raise HTTPException(status_code=401, detail='Authentication Failed')
    return db.query(Todos).all()


@router.delete('/delete',status_code=status.HTTP_204_NO_CONTENT)
async def delete_todo(user: user_dependency, db:db_dependency,todo_id: int):
    if user is None or user.get('user_role') != 'admin':
        raise HTTPException(status_code=401, detail='Authentication Failed')
    todo_delete = db.query(Todos).filter(Todos.id==todo_id).first()
    if todo_delete is not None:
        db.query(Todos).filter(Todos.id == todo_id).delete()
        db.commit()
